Bug 427003 Upload of binary files with invalid UTF-8 byte sequences
FileHandlerV1#handlePutContents() used an UTF-8 encoded reader
for copying the request entities to workspace files.
This in general does not work when the input contains invalid
UTF-8 byte sequences, because readers tend to replace the invalid
byte sequences with replacement characters like \uFFFD. 

This patch removes the usage of UTF-8 encoded readers for the body
part, but not for the metadata in multipart requests!

Change-Id: I4db00623bb4fa5044c22b77559550b1c01d81781
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>