Bug 427666 Fixed xfer with multipart content type
Content-Length is always larger than actual payload of a multipart
request because of the necessary leading and trailing delimiter
strings (and optional headers between leading delimiter and body of
the part).
ClientImport however, enforced equality of Content-Length and
given byte range, which always failed for multipart requests.

This patch replaces that check with one that enforce equality of
the actual chunk size read from the request stream with the
announced byte range. This check works well for multipart and
non-multipart requests.

Change-Id: Iba5f282ab658f47d784676b24af4191c6661cb05
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>