Bug 427667 Removed the buffering of chunks in ClientImport
ClientImport buffers complete chunks in byte arrays before writing
them to the file system without applying a suitable quota; even
worse, in the multipart case, the chunk is buffered in a StringBuffer
taking twice the chunk size of heap.

This patch removes the buffering in the multipart and non-multipart
cases and replaces it with a channeling solution based on
FileChannel#transferFrom().

Change-Id: Icab2ebc0e43a359c3d7da4ff676e3e538e198465
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>