Bug 428733 Clone should fall back to GitName/GitMail from user prefs
The old implementation relied on the fact that the client sends
GitName/GitMail parameters with the clone command --- which it before
obtained from the server with a /prefs/user call. However, if the
client fails to send these parameters, the server should fall back
to the parameters maintained in the preferences.

Change-Id: Idf2c79fbcf8bc1b2d61929f605874e44bedeff29
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>