Bug 442727 - GET/PUT/DELETE on /prefs/user should be forbidden
This patch adds the missing path segment checks for the GET,
PUT and DELETE requests, so that requests to /prefs/user,
/prefs/workspace and /prefs/project are now rejected with a
405 Method Not Allowed.

Furthermore, this patch adds a check that prevents prefixes starting
with the string "User". All critical preferences, i.e. "UserRights"
and "UserQuota" fall into that category and judging from the original
code these properties were never meant to  be accessible via the
preferences API. Properties starting with the prefix "User" are now
also removed from the responses of GET requests.

Change-Id: Iecf2d6a7f79f6c9e2347af898064ad06311070cb
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>