Bug 435067 - Adding XSRFPreventionFilter
The filter implements the "Double Submit Cookies" strategy [1] to
protect Orion against cross site request forgery attacks. It creates a
new nonce and sets it as cookie when a user is authenticated.

The filter checks for all requests except 'GET'-requests if a valid
nonce is provided in the request header and compares it with the value
from the cookie. If both values/nonces are not the same a "403
FORBIDDEN" response is sent.

Further a ServerConstant is set to enable disabling of the XSRF
protection filter.

This change depends on change Ica5b002300c200f3849df579e4c6a4859e15ef07
in org.eclipse.orion.client.

[1]
https://www.owasp.org/index.php/Cross-Site_Request_Forgery_(CSRF)_Prevention_Cheat_Sheet#Double_Submit_Cookies

Change-Id: I849eb62c46ab52b59e3008d3a998d382310ae91c
Signed-off-by: Erwin Margewitsch <erwin.margewitsch@sap.com>