Bug 435067 - Adding XSRFPreventionFilter
The filter creates a new nonce and sets it as cookie when a user is
authenticated.

The filter checks for all requests except 'GET'-requests if a valid
nonce is provided in the request header and compares it with the value
from the cookie. If both values/nonces are not the same a "403
FORBIDDEN" response is sent.

Change-Id: I849eb62c46ab52b59e3008d3a998d382310ae91c
Signed-off-by: Erwin Margewitsch <erwin.margewitsch@sap.com>