Bug 435067 - Adding XSRFPreventionFilter
The filter creates a new nonce or returns the existing one if the header
contains the xsrf-token with value 'fetch'. The nonce is stored as
property in the session of the request.

The filter checks for all requests except 'GET'-requests if a valid
nonce is provided in the request header. If the request doesn't contain
a valid nonce a "403 FORBIDDEN" response is sent.

Change-Id: I849eb62c46ab52b59e3008d3a998d382310ae91c
Signed-off-by: Erwin Margewitsch <erwin.margewitsch@sap.com>