Bug 443181 - Missing diff in log response for file creation
JGit sometimes initializes RevCommits only partially for performance
reasons, and one has to explicitly retrieve the missing attributes
with a RevWalk if needed. This was done in Commit#getDiffs() for
the parent of the given RevCommit, but not for the RevCommit itself.

This patch adds a RevWalk#parseCommit() call in the constructor of
the Commit data object, which is called in case the "parents"
collection of the passed in RevCommit is empty.

Change-Id: Iece7ff2a6585ca3093fdbe298907e38d6fc250e9
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>