Bug 443808 - PushJob should deliver backend messages to client
Some servers like Gerrit deliver additional information in the
response of a push request, e.g. links to created or updated
changes, which may be useful for a developer.

This patch converts any message attached to the returned JGit
PushResult into a "DetailedMessage" attribute of the corresponding
Orion push response.

Example with a detail message from a Gerrit server:
{
  "Location": "/task/id/EMr7rmVNABQQAdFXjLxe_w",
  "Result": {
    "BundleId": "unknown",
    "Code": 0,
    "DetailedMessage": "\r
      Processing changes: updated: 1, refs: 1\r
      Processing changes: updated: 1, refs: 1, done    \n\n
      Updated Changes:\n
        https://git.example.org/3762\n\n",
    "HttpCode": 200,
    "Message": "OK",
    "Severity": "Ok"
  },
  ...
}

Change-Id: I601a5b78f66c548ab362a5bca5a61d044051c47c
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>